factor out analyte name helpers, analytes_sorted() method
