+++ latools/helpers/helpers.py
@@ -132,69 +132,6 @@ def unitpicker(a, llim=0.1, denominator=None, focus_stage=None):
             n += 1
     return float(1000**n), udict[n]
 
-def pretty_element(s):
-    """
-    Returns formatted element name.
-
-    Parameters
-    ----------
-    s : str
-        of format [A-Z][a-z]?[0-9]+
-
-    Returns
-    -------
-    str
-        LaTeX formatted string with superscript numbers.
-    """
-    el = re.match('.*?([A-z]{1,3}).*?', s).groups()[0]
-    m = re.match('.*?([0-9]{1,3}).*?', s).groups()[0]
-
-    return '$^{' + m + '}$' + el
-
-def get_analyte_name(s):
-    return re.match('.*?([A-z]{1,3}).*?', s).groups()[0]
-
-def get_analyte_mass(s):
-    return re.match('.*?([0-9]{1,3}).*?', s).groups()[0]
-
-def analyte_2_namemass(s):
-    """
-    Converts analytes in format '27Al' to 'Al27'.
-
-    Parameters
-    ----------
-    s : str
-        of format [A-z]{1,3}[0-9]{1,3}
-
-    Returns
-    -------
-    str
-        Name in format [0-9]{1,3}[A-z]{1,3}
-    """
-    el = re.match('.*?([A-z]{1,3}).*?', s).groups()[0]
-    m = re.match('.*?([0-9]{1,3}).*?', s).groups()[0]
-
-    return el + m
-
-def analyte_2_massname(s):
-    """
-    Converts analytes in format 'Al27' to '27Al'.
-
-    Parameters
-    ----------
-    s : str
-        of format [0-9]{1,3}[A-z]{1,3}
-
-    Returns
-    -------
-    str
-        Name in format [A-z]{1,3}[0-9]{1,3}
-    """
-    el = re.match('.*?([A-z]{1,3}).*?', s).groups()[0]
-    m = re.match('.*?([0-9]{1,3}).*?', s).groups()[0]
-
-    return m + el
-
 def collate_data(in_dir, extension='.csv', out_dir=None):
     """
     Copy all csvs in nested directroy to single directory.
