+++ latools/preprocessing/split.py
@@ -13,7 +13,8 @@ import pandas as pd
 import pkg_resources as pkgrs
 from warnings import warn
 from ..processes import read_data, autorange
-from ..helpers.helpers import analyte_2_namemass, bool_2_indices
+from ..helpers.helpers import bool_2_indices
+from ..helpers.analyte_names import analyte_2_namemass
 from ..helpers.io import read_dataformat
 
 import matplotlib.pyplot as plt
